Disease Identification Model/requirements.txt
Flask
Flask-Cors
tensorflow
pandas
numpy
huggingface_hub
python-dotenv
opencv-python
